# requisitos/admin/ej-admin-openapi.yaml
openapi: 3.0.3
info:
  title: Elastic Journey Admin Portal API
  description: >
    API do Elastic Journey Admin Portal para gestão de produtos (e dos tipos de
    canal que cada um habilita), jornadas multicanal, fluxos, formulários,
    execuções e publicações. A publicação aciona uma API externa mockada na
    versão 1.0.0; este contrato descreve somente a API do Admin Portal.
  version: 1.0.0
servers:
  - url: https://admin.elasticjourney.com/api/v1
    description: Production
  - url: https://admin-hml.elasticjourney.com/api/v1
    description: Homologation

# Na versão 1.0.0, a autenticação utiliza um provedor externo mockado. O acesso inicial
# usa o usuário admin, senha admin e papel ADMIN; as operações devem preservar
# a identidade autenticada e aplicar os papéis ADMIN, EDITOR e VIEWER.
security:
  - bearerAuth: []

# Observabilidade (FT-10): toda requisição aceita o header opcional
# X-Correlation-Id; quando ausente, o servidor gera um valor. O mesmo valor é
# devolvido no header de resposta X-Correlation-Id e correlaciona, nos logs
# técnicos do backend, a entrada/saída da requisição às transações de
# persistência por ela disparadas. Cross-cutting, omitido dos schemas de
# request/response de cada operação abaixo para não repetir o mesmo header em
# cada endpoint.
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products
      operationId: listProducts
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE]
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Products]
      summary: Create product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}:
    get:
      tags: [Products]
      summary: Get product
      operationId: getProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
    put:
      tags: [Products]
      summary: Update product
      operationId: updateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  /products/{productId}/deactivate:
    post:
      tags: [Products]
      summary: Deactivate product
      description: Returns 409 while any journey under the product has a PUBLISHED publication.
      operationId: deactivateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Product deactivated
  /products/{productId}/activate:
    post:
      tags: [Products]
      summary: Reactivate product
      operationId: activateProduct
      parameters:
        - $ref: '#/components/parameters/ProductId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Product reactivated
  /journeys:
    get:
      tags: [Journeys]
      summary: List journeys
      operationId: listJourneys
      parameters:
        - name: productId
          in: query
          schema:
            type: string
            format: uuid
        - name: channelType
          in: query
          schema:
            $ref: '#/components/schemas/ChannelType'
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          description: Sort order, most recent first
          schema:
            type: string
            enum: [CREATED_AT, UPDATED_AT]
            default: UPDATED_AT
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Journey list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Journey'
    post:
      tags: [Journeys]
      summary: Create journey
      operationId: createJourney
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JourneyCreateInput'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '201':
          description: Journey created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Journey'
...
  /journeys/{journeyId}:
    get:
      tags: [Journeys]
      summary: Get journey
      operationId: getJourney
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Journey found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Journey'
    put:
      tags: [Journeys]
      summary: Update journey
      operationId: updateJourney
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JourneyUpdateInput'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Journey updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Journey'
    delete:
      tags: [Journeys]
      summary: Delete a journey — physically if never published, soft-deleted (INACTIVE) otherwise
      description: Returns 409 if the journey is currently PUBLISHED or already INACTIVE.
      operationId: deleteJourney
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '204':
          description: Journey deleted
        '409':
          $ref: '#/components/responses/Conflict'
  /journeys/{journeyId}/flow:
    get:
      tags: [Flows]
      summary: Get journey flow
      operationId: getFlow
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Flow found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
    put:
      tags: [Flows]
      summary: Update journey flow
      description: >
        Replaces the complete flow. The backend returns 422 unless the payload
        contains exactly one initial element (START or MESSAGE_START_EVENT),
        exactly one END and every node belongs to a continuous path between them.
      operationId: updateFlow
      parameters:
        - $ref: '#/components/parameters/JourneyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Flow'
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '200':
          description: Flow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Flow'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: mock-session
  responses:
    BadRequest:
      description: Malformed request or invalid parameter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Missing or unauthenticated identity; reserved for future use
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Forbidden:
      description: Authenticated identity is not allowed to perform the operation; reserved for future use
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: Request conflicts with the current resource state or a uniqueness constraint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnprocessableEntity:
      description: Request is syntactically valid but violates a functional rule
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalServerError:
      description: Unexpected internal failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  parameters:
    ProductId:
      name: productId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    JourneyId:
      name: journeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ApiError:
      type: object
      required: [timestamp, status, code, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        code:
          type: string
          description: Stable machine-readable error identifier
        message:
          type: string
          description: Human-readable error description
        path:
          type: string
          description: Request path that produced the error
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiErrorDetail'
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    ProductInput:
      type: object
      required: [name, description, channelTypes]
      properties:
        name:
          type: string
          maxLength: 150
        description:
          type: string
        channelTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ChannelType'
    Product:
      allOf:
        - $ref: '#/components/schemas/ProductInput'
        - type: object
          required: [productId, status, createdAt, updatedAt]
          properties:
            productId:
              type: string
              format: uuid
            status:
              type: string
              enum: [ACTIVE, INACTIVE]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # Valor de domínio fixo — não é uma entidade cadastrável, não tem CRUD nem status próprio.
    # Substitui a antiga entidade Channel (channelId/name/status por produto, 6 tipos incluindo
    # URA/CONTACT_CENTER/OTHER) — ver ej-admin-requisitos.md FT-01.
    ChannelType:
      type: string
      enum: [WEB, MOBILE, WHATSAPP]
    JourneyCreateInput:
      type: object
      required: [productId, channelTypes, name, description]
      properties:
        productId:
          type: string
          format: uuid
        channelTypes:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ChannelType'
        name:
          type: string
          maxLength: 200
        description:
          type: string
        templateId:
          type: string
          maxLength: 100
          description: Stable id from GET /journey-templates. Omit to create a blank flow.
...
    JourneyUpdateInput:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
          maxLength: 200
        description:
          type: string
    Journey:
      allOf:
        - $ref: '#/components/schemas/JourneyUpdateInput'
        - type: object
          required: [journeyId, channelTypes, productId, productName, status, createdAt, updatedAt]
          properties:
            journeyId:
              type: string
              format: uuid
            channelTypes:
              type: array
              minItems: 1
              uniqueItems: true
              items:
                $ref: '#/components/schemas/ChannelType'
            productId:
              type: string
              format: uuid
            productName:
              type: string
            status:
              type: string
              enum: [DRAFT, PUBLISHED, UNPUBLISHED, INACTIVE]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    Flow:
      type: object
      description: >
        Persisted as a single JSONB document per journey (flow, journeyId,
        name, nodes, connections, annotations). flowId, FlowNode.nodeId and
        FlowConnection.connectionId are the identifiers later embedded
        verbatim as BPMN element ids by the runtime when the flow is
        translated to a runtime engine process definition; XML ID/NCName forbids an
        identifier starting with a digit, which a raw UUID cannot guarantee.
        These three identifiers are therefore generated as a fixed prefix
        + UUID (Process_<uuid>, Node_<uuid>, Flow_<uuid> respectively), never
        a bare UUID. All other entity identifiers in this API (productId,
        journeyId, formId, ...) never appear inside generated BPMN
        XML and remain plain UUIDs. annotations are canvas-only notes, never
        translated to BPMN and never embedded with a runtime-facing id.
      required: [flowId, journeyId, name, nodes, connections, annotations]
      properties:
        flowId:
          type: string
          example: Process_f47ac10b-58cc-4372-a567-0e02b2c3d479
        journeyId:
          type: string
          format: uuid
        name:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/FlowNode'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/FlowConnection'
        annotations:
          type: array
          items:
            $ref: '#/components/schemas/FlowAnnotation'
    FlowNode:
      type: object
      required: [nodeId, nodeType, name]
      properties:
        nodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
        nodeType:
          type: string
          enum: [START, END, USER_TASK, SERVICE_TASK, RECEIVE_TASK, MESSAGE_START_EVENT]
        name:
          type: string
        description:
          type: string
        positionX:
          type: integer
        positionY:
          type: integer
        userTaskConfig:
          $ref: '#/components/schemas/UserTaskConfig'
        connectorConfig:
          $ref: '#/components/schemas/IntegrationTaskConfig'

    # Nota de revisão (2026-09-05): embeddedScreen (lista de FormField) substituído por
    # embeddedScreenRoot (SduiNode único) — catálogo SDUI corporativo v1, ver ej-admin-requisitos.md
    # FT-04. A árvore editada aqui é a mesma árvore publicada, sem etapa de compilação.
    UserTaskConfig:
      type: object
      properties:
        embeddedScreenRoot:
          description: >
            Raiz da árvore de tela (SduiNode) desenhada diretamente no nó — ausente quando a
            User Task não tem tela (nesse caso, messageText é usado). Nunca coexiste com
            sentido junto de messageText. Sempre um único nó do tipo ui.screen.
          $ref: '#/components/schemas/SduiNode'
          nullable: true
        messageText:
          type: string
    IntegrationTaskConfig:
      type: object
      description: >
        Exposed as FlowNode.connectorConfig — not a standalone resource:
        connectorType/config/credentialRef are attributes of the FlowNode
        item itself (Flow.nodes, JSONB array). There is no connectorEnabled
        per node: whether a connector type is available at all is a
        catalog-level filter on the type list offered to the user, not a
        stored field. `config` is a free-form map, not a discriminated
        union — RestConnectorConfig/KafkaConnectorConfig below document its
        real shape per connectorType for reference, they are not separate
        schema types on the wire.
      required: [connectorType, config]
      properties:
        connectorType:
          type: string
          enum: [REST, KAFKA, EVENT_HUBS, SERVICE_BUS]
        config:
          type: object
          additionalProperties: true
          description: >
            REST: see RestConnectorConfig. KAFKA/EVENT_HUBS/SERVICE_BUS: see
            KafkaConnectorConfig. outputMapping lives inside this map for
            both — there is no separate top-level outputMapping field, and
            no inputMapping field at all (text fields like url/headers/body/
            payload just reference prior-step variables inline via
            `{{name}}`, REQ-03.09.012).
        credentialRef:
          type: string
          maxLength: 200
          description: Runtime-resolved credential reference; secret values are not stored.
...
    FlowConnection:
      type: object
      required: [connectionId, sourceNodeId, targetNodeId]
      properties:
        connectionId:
          type: string
          example: Flow_f47ac10b-58cc-4372-a567-0e02b2c3d479
        sourceNodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
        targetNodeId:
          type: string
          example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
    FlowAnnotation:
      type: object
      description: >
        Free-floating note on the flow canvas, documentation-only — never
        validated by FlowValidator, never translated to BPMN. id is a plain
        Annotation_<uuid> by convention, never embedded in generated BPMN
        XML, so it carries no NCName constraint.
      required: [id, text, positionX, positionY]
      properties:
        id:
          type: string
          example: Annotation_f47ac10b-58cc-4372-a567-0e02b2c3d479
        text:
          type: string
        positionX:
          type: integer
        positionY:
          type: integer
        linkedNodeIds:
          type: array
          items:
            type: string
            example: Node_f47ac10b-58cc-4372-a567-0e02b2c3d479
          description: FlowNode ids in the same flow this note is visually tied to (dashed line in the editor). Informational only.

    # FormField (modelo de campo plano do catálogo SDUI anterior) removido em 2026-09-05, junto com
    # US-04.02 — ver ej-admin-requisitos.md FT-04. Substituído por SduiNode (árvore recursiva) +
    # ComponentDefinition (Component Registry, schema mais abaixo).
...